Add tests for JPEG2K
